add token to cycler
